still trying to figure out issues with installing from npm...

+++ bootstrap.js
@@ -8,18 +8,25 @@ var util      = require('./src/util'),
     shell     = require('shelljs'),
     platforms = require('./platforms');
 
-// Create native projects using bin/create
-var tempDir = path.join(__dirname, 'spec', 'fixtures', 'projects', 'native');
-shell.rm('-rf', tempDir);
-shell.mkdir('-p', tempDir);
+var cmd = 'android update project -p ' + path.join(__dirname, 'lib', 'cordova-android', 'framework') + ' -t android-17';
+shell.exec(cmd, {async:true}, function(code, output) {
+    if (code > 0) {
+        process.exit(1);
+    } else {
+        // Create native projects using bin/create
+        var tempDir = path.join(__dirname, 'spec', 'fixtures', 'projects', 'native');
+        shell.rm('-rf', tempDir);
+        shell.mkdir('-p', tempDir);
 
-platforms.forEach(function(platform) {
-    var fix_path = path.join(tempDir, platform + '_fixture');
-    var create = path.join(util.libDirectory, 'cordova-' + platform, 'bin', 'create'); 
-    console.log('Creating cordova-' + platform + ' project using live project lib for tests...');
-    var cmd = create + ' "' + fix_path + '" org.apache.cordova.cordovaExample cordovaExample';
-    if (platform == 'blackberry') cmd = create + ' "' + fix_path + '" cordovaExample';
-    var create_result = shell.exec(cmd, {silent:true});
-    if (create_result.code > 0) throw ('Could not create a native ' + platform + ' project test fixture: ' + create_result.output);
-    console.log('.. complete.');
+        platforms.forEach(function(platform) {
+            var fix_path = path.join(tempDir, platform + '_fixture');
+            var create = path.join(util.libDirectory, 'cordova-' + platform, 'bin', 'create'); 
+            console.log('Creating cordova-' + platform + ' project using live project lib for tests...');
+            var cmd = create + ' "' + fix_path + '" org.apache.cordova.cordovaExample cordovaExample';
+            if (platform == 'blackberry') cmd = create + ' "' + fix_path + '" cordovaExample';
+            var create_result = shell.exec(cmd, {silent:true});
+            if (create_result.code > 0) throw ('Could not create a native ' + platform + ' project test fixture: ' + create_result.output);
+            console.log('.. complete.');
+        });
+    }
 });
